Iam auth

+++ controllers/experimentController.js
@@ -1,41 +1,75 @@
 const axios = require('axios');
 const ExperimentRepository = require('../repositories/experimentRepository');
 const experimentRepository = new ExperimentRepository();
+const UserRepository = require('../repositories/userRepository');
+const userRepository = new UserRepository();
+
+function checkManagerAuth(details) {
+    if(details.type != "manager" || details.credits <= 0) {
+        return false;
+    } else {
+        return true;
+    }
+}
 
 exports.experimentController = {
-    createExperiment(req, res) {
-        // axios.get('https://iam-shenkar.onrender.com/assets/token')
-        //     .then(response => {
-        //         const user = response.data;
-        //         if (user.type !== 'manager')
-        //             res.send("Not Authorised"); // should change to page
-        //     })
-        //     .catch(mock => {
-        //         const user = IAMRepository.getUserById(id);
-        //         res.send(experimentRepository.getExperimentById(id));
-        //     })
+    async createExperiment(req, res) {
+        let details;
+        //getting type, credits, plan assets from IAM
+        await axios.get('https://am-shenkar.onrender.com/assets', {headers: {'Content-Type': 'application/json'}})
+            .then(response => {
+                console.log(response.data);
+                details = response.data;
+            })
+            .catch(mock => {
+                details = userRepository.getDetailsById(req.session.userId);
+            })
 
-        axios.post('https://growth.render.com/experiment/new', req.body)
+        if (checkManagerAuth(details)) {
+            await axios.post('https://growth.render.com/experiment/new', req.body)
+                .then(async response => {
+                    await axios.put('https://am-shenkar.onrender.com/credits/1', {headers: {'Content-Type': 'application/json'}})
+                        .then(response => {
+                            res.send(response.data);
+                        })
+                        .catch(mock => {
+                            userRepository.decreaseCredit(req.session.userId, 1);
+                            const id = experimentRepository.createExperiment(req.body);
+                            res.send(experimentRepository.getExperimentById(id));
+                        })
+                })
+                .catch(mock => {
+                    const id = experimentRepository.createExperiment(req.body);
+                    res.send(experimentRepository.getExperimentById(id));
+                })
+        } else {
+            res.send("you don't have permission to create experiment")
+        }
+
+    },
+    async updateExperiment(req, res) {
+        let details;
+        //getting type, credits, plan assets from IAM
+        await axios.get('https://am-shenkar.onrender.com/assets', {headers: {'Content-Type': 'application/json'}})
             .then(response => {
-                res.send(response.data);
+                details = response.data;
             })
             .catch(mock => {
-                const id = experimentRepository.createExperiment(req.body);
-                res.send(experimentRepository.getExperimentById(id));
+                details = userRepository.getDetailsById(req.session.userId);
             })
+        if (checkManagerAuth(details)) {
+            await axios.put(`https://growth.render.com/experiment/${req.params.id}`, req.body)
+                .then(response => {
+                    res.send(response.data);
+                })
+                .catch(mock => {
+                    const updatedExperiment = experimentRepository.updateExperiment(req.body, req.params.id);
+                    res.send(updatedExperiment);
+                })
+        }
     },
-    updateExperiment(req, res) {
-        axios.put(`https://growth.render.com/experiment/${req.params.id}`, req.body)
-            .then(response => {
-                res.send(response.data);
-            })
-            .catch(mock => {
-                const updatedExperiment = experimentRepository.updateExperiment(req.body, req.params.id);
-                res.send(updatedExperiment);
-            })
-    },
-    experimentStatistics(req , res) {
-        axios.get(`https://growth.render.com/experiment/${req.params.id}/statistics`)
+    async experimentStatistics(req , res) {
+        await axios.get(`https://growth.render.com/experiment/${req.params.id}/statistics`)
             .then(response => {
                 res.send(response.data);
             })
@@ -43,8 +77,8 @@ exports.experimentController = {
                 res.send(`Statistics data from experiment ${req.params.id}`);
             })
     },
-    endExperiment(req, res) {
-        axios.post(`https://growth.render.com/experiment/${req.params.id}/statistics`, {
+    async endExperiment(req, res) {
+        await axios.post(`https://growth.render.com/experiment/${req.params.id}/statistics`, {
             experimentId: req.body.experimentId
         })
             .then(response => {
@@ -54,18 +88,34 @@ exports.experimentController = {
                 res.send(`experiment ${req.params.id} ended`);
             })
     },
-    experimentsByAccount(req, res) {
-        axios.get(`https://growth.render.com/experiment/${req.params.account}`)
+    async experimentsByAccount(req, res) {
+        let status;
+        await axios.get('https://am-shenkar.onrender.com/assets/token', {headers: {'Content-Type': 'application/json'}})
             .then(response => {
-                res.send(response.data);
+                if('message' in response.data)
+                    status = true;
+                else
+                    status = false;
             })
             .catch(mock => {
-                const data = experimentRepository.getExperimentByAccount(req.params.account);
-                res.send(data);
+                if('userId' in req.session)
+                    status = true;
+                else
+                    status = false;
             })
+        if(status) {
+            await axios.get(`https://growth.render.com/experiment/${req.params.account}`)
+                .then(response => {
+                    res.send(response.data);
+                })
+                .catch(mock => {
+                    const data = experimentRepository.getExperimentByAccount(req.params.account);
+                    res.send(data);
+                })
+        }
     },
-    ABTestByAccount(req, res) {
-        axios.get(`https://growth.render.com/experiment/AB/${req.params.account}`)
+    async ABTestByAccount(req, res) {
+        await axios.get(`https://growth.render.com/experiment/AB/${req.params.account}`)
             .then(response => {
                 res.send(response.data);
             })
@@ -74,8 +124,8 @@ exports.experimentController = {
                 res.send(data);
             })
     },
-    FeatureFlagByAccount(req, res) {
-        axios.get(`https://growth.render.com/experiment/FF/${req.params.account}`)
+    async FeatureFlagByAccount(req, res) {
+        await axios.get(`https://growth.render.com/experiment/FF/${req.params.account}`)
             .then(response => {
                 res.send(response.data);
             })
@@ -84,18 +134,29 @@ exports.experimentController = {
                 res.send(data);
             })
     },
-    deleteExperiment(req, res) {
-        axios.delete(`https://growth.render.com/experiment/${req.params.id}`)
+    async deleteExperiment(req, res) {
+        let details;
+        //getting type, credits, plan assets from IAM
+        await axios.get('https://am-shenkar.onrender.com/assets', {headers: {'Content-Type': 'application/json'}})
             .then(response => {
-                res.send(response.data);
+                details = response.data;
             })
             .catch(mock => {
-                experimentRepository.deleteExperiment(req.params.id);
-                res.send(`experiment ${req.params.id} deleted`);
+                details = userRepository.getDetailsById(req.session.userId);
             })
+        if (checkManagerAuth(details)) {
+            await axios.delete(`https://growth.render.com/experiment/${req.params.id}`)
+                .then(response => {
+                    res.send(response.data);
+                })
+                .catch(mock => {
+                    experimentRepository.deleteExperiment(req.params.id);
+                    res.send(`experiment ${req.params.id} deleted`);
+                })
+        }
     },
-    callExperiment(req, res) {
-        axios.post(`https://growth.render.com/experiment/${req.params.id}`, req)
+    async callExperiment(req, res) {
+        await axios.post(`https://growth.render.com/experiment/${req.params.id}`, req)
             .then(response => {
                 res.send(response.data);
             })
@@ -109,8 +170,8 @@ exports.experimentController = {
 
             })
     },
-    declareGoal(req, res) {
-        axios.post(`https://growth.render.com/experiment/goal/${req.params.id}`, req)
+    async declareGoal(req, res) {
+        await axios.post(`https://growth.render.com/experiment/goal/${req.params.id}`, req)
             .then(response => {
                 res.send("declared goal");
             })
